split sed into 32 and 64 bit versions
These versions are identical except for the name of the exe they create.
I did this just to make it so I can run the seds from iexpress, instead
of editing and clicking next until I get to the "name the exe" screen.
These should be run by the appropriate iexpress - syswow64 for 32bit and
system32 for 64 bit...