fixed installer and install script
The installer wasn't copying the virtualenv.bat shim, so I fixed it.

The install command wasn't handling installing with certs correctly.  I
removed the cases where you could mix the --no-check-certificate and
--ca-certificate flags, because mixing them is pointless.  I changed
from --certificate-directory to --ca-certificate because I realized the
cert format required for the directory is complex to obtain in windows.
--ca-certificate requires a bundle of master certs a la what you can get
from mozilla.  I added instructions to the readme about getting certs
from the curl folks

Finally, I decided there is no point to a 64 bit installer, so I'm just
not going to make one